make index.js not one huge file
