+++ backend/middleware/auth.js
@@ -1,0 +1,21 @@
+import jwt from 'jsonwebtoken';
+
+// JWT secret - in production, use environment variable
+const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key-change-in-production';
+
+export const verifyToken = async (request, reply) => {
+  try {
+    const token = request.cookies.token || 
+                  request.headers.authorization?.replace('Bearer ', '');
+    if (!token) {
+      return reply.code(401).send({ error: 'No token provided' });
+    }
+
+    const decoded = jwt.verify(token, JWT_SECRET);
+    request.user = decoded;
+  } catch (err) {
+    return reply.code(401).send({ error: 'Invalid token' });
+  }
+};
+
+export { JWT_SECRET };
