+++ backend/package.json
@@ -8,11 +8,10 @@
     "dev": "nodemon index.js"
   },
   "dependencies": {
-    "fastify": "^4.24.3",
-    "@fastify/cors": "^8.4.0",
-    "@fastify/cookie": "^9.2.0",
-    "mysql2": "^3.6.5",
-    "bcrypt": "^5.1.1",
+    "@fastify/cookie": "^11.0.2",
+    "@fastify/cors": "^11.0.1",
+    "@fastify/mysql": "^5.0.2",
+    "fastify": "^5.4.0",
     "jsonwebtoken": "^9.0.2"
   },
   "devDependencies": {
